Made a lot of ScopedPointer usage conform to the std::unique_ptr interface

+++ extras/NetworkGraphicsDemo/Source/MasterComponent.h
@@ -42,8 +42,8 @@ struct MasterContentComponent  : public Component,
         createAllDemos (demos);
         setContent (0);
 
-        setSize (15.0f * currentCanvas.getLimits().getWidth(),
-                 15.0f * currentCanvas.getLimits().getHeight());
+        setSize ((int) (15.0f * currentCanvas.getLimits().getWidth()),
+                 (int) (15.0f * currentCanvas.getLimits().getHeight()));
 
         if (! OSCSender::connect (getBroadcastIPAddress(), masterPortNumber))
             error = "Master app OSC sender: network connection error.";
